+++ TrailAid.Models/Trail/TrailCreate.cs
@@ -35,7 +35,7 @@ namespace TrailAid.Models.Trail
         /// <summary>
         /// trail distance
         /// </summary>
-        public int Distance { get; set; }
+        public double Distance { get; set; }
         /// <summary>
         /// terrain type
         /// </summary>
@@ -47,7 +47,7 @@ namespace TrailAid.Models.Trail
         /// <summary>
         /// trail elevation
         /// </summary>
-        public int Elevation { get; set; }
+        public double Elevation { get; set; }
         /// <summary>
         /// trail route type
         /// </summary>
